moved DTOs to request & response folder,  made PostAnnouncementReq for createAnnouncement API, added annotations for swagger to controllers (@Api, @ApiOperation)
