+++ Swith-Server/src/main/java/com/example/demo/src/controller/GroupInfoController.java
@@ -2,13 +2,14 @@ package com.example.demo.src.controller;
 
 import com.example.demo.config.BaseException;
 import com.example.demo.config.BaseResponse;
-import com.example.demo.src.dto.GetGroupInfoRes;
-import com.example.demo.src.dto.GetHomeGroupInfoRes;
+import com.example.demo.src.dto.response.GetHomeGroupInfoRes;
 import com.example.demo.src.dto.PostGroupInfoReq;
 import com.example.demo.src.dto.PostGroupInfoRes;
 import com.example.demo.src.service.GroupInfoService;
 import com.example.demo.src.service.SessionService;
 import com.example.demo.src.service.UserService;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,6 +17,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/groupinfo")
+@Api(tags = {"Swith GroupInfo API"})
 public class GroupInfoController {
 
     private final GroupInfoService groupInfoService;
@@ -27,15 +29,15 @@ public class GroupInfoController {
         this.userService = userService1;
     }
 
-    @ResponseBody
+    @ApiOperation("홈화면 정보 불러오기")
     @GetMapping("/home")
     public BaseResponse<List<GetHomeGroupInfoRes>> loadHomeData (@RequestParam(value = "userIdx") Long userIdx) throws BaseException {
         List<GetHomeGroupInfoRes> getGroupHomeData = groupInfoService.loadHomeData(userIdx);    //출석율 부분 수정 필요
         return new BaseResponse<>(getGroupHomeData);
     }
 
-    @PostMapping("")
-    @ResponseBody
+    @ApiOperation("그룹 생성")
+    @PostMapping
     public BaseResponse<PostGroupInfoRes> createGroup(@RequestBody PostGroupInfoReq request){
         System.out.println(request.toString());
         System.out.println("들어가기전!");
