+++ Swith-Server/src/main/java/com/example/demo/src/dto/response/GetGroupInfoRes.java
@@ -1,5 +1,6 @@
-package com.example.demo.src.dto;
+package com.example.demo.src.dto.response;
 
+import com.example.demo.src.dto.response.GetSessionRes;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.*;
